Create a ToolFunctionDefinition struct
